+++ src/main/c/cext/numeric.c
@@ -1,4 +1,5 @@
 #include <truffleruby-impl.h>
+#include <ruby/util.h>
 
 // Numeric conversions, rb_*num*, rb_fix2*
 
@@ -25,6 +26,13 @@ void rb_num_zerodiv(void) {
 }
 
 // Conversions between numeric types and from/to String
+
+double ruby_strtod(const char *nptr, char **endptr) {
+  // See ruby/util.h
+  #undef strtod
+  return strtod(nptr, endptr);
+  #define strtod(s,e) ruby_strtod((s),(e))
+}
 
 unsigned long rb_num2ulong(VALUE val) {
   return (unsigned long)polyglot_as_i64(RUBY_CEXT_INVOKE_NO_WRAP("rb_num2ulong", val));